+++ src/main/java/com/lne_wizards/LNE_Wizards_Mod.java
@@ -22,7 +22,7 @@ public class LNE_Wizards_Mod implements ModInitializer {
 			.sanitize(true)
 			.build();
 	public static ConfigManager<TweaksConfig> tweaksConfig = new ConfigManager<>
-			("tweaks", new TweaksConfig())
+			("tweaks_v1", new TweaksConfig())
 			.builder()
 			.setDirectory(MOD_ID)
 			.sanitize(true)