+++ src/main/java/com/lne_wizards/config/TweaksConfig.java
@@ -3,9 +3,9 @@ package com.lne_wizards.config;
 public class TweaksConfig {
     public boolean disable_special_lne_weapons = false;
     public int everfrost_staff_freeze_ticks_on_damage = 5;
-    public int everfrost_staff_freezing_effect_seconds = 2;
+    public int everfrost_staff_freezing_effect_seconds = 5;
     public int netherflame_staff_haste_max_amplifier = 3;
-    public int netherflame_staff_haste_duration = 4;
+    public int netherflame_staff_haste_duration = 8;
     public int dragon_staff_arcane_precision_max_amplifier = 5;
     public int dragon_staff_arcane_precision_duration = 10;
 }